closes #25; go build succeeds
