+++ main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/softsrv/gamify/internal/pkg/router"
+	"github.com/softsrv/gamify/pkg/router"
 )
 
 func main() {
